Clarify which registry is used for the result of create*Document()
Documents don't have associated documents, so "the associated document to the new instance" didn't pick out any particular document. `DOMImplementation`s are associated with documents, and I believe that's the document that was meant here. (Another option would be the global object’s associated document.) The DOM spec tends to use "the associated document of the context object" in methods of `DOMImplementation`, but I think the more direct form I'm using here is easier to read.