Update custom element lifecycle
This redefines the custom element callback queue to be a custom element
action queue, to allow upgrades (which are no longer callback-based) to
take part. It also replaces attachedCallback/detachedCallback with
connectedCallback/disconnectedCallback, fixing #362. And it updates all
the places that enqueue these callbacks to be phrased in terms of
precise patches to DOM, except for upgrades.