Overhaul custom element upgrading to be much more explicit
This eliminates a lot of action-at-a-distance from the spec, in favor of
having the appropriate steps set or unset the new "defined flag". This
fixes #425. We also make the element interface concept more explicit,
with a new section stating the exact text that should be added to HTML.

While doing this, also  changed :unresolved to :defined, which fixes
#418. You can use `:not(:defined)` for the effect of the old
`:unresolved`.

The term "resolved" has been eliminated from the spec in favor of
"defined".