Remove base element queue and related machinery
Per #457, it is used for async parsing/construction, but the spec no longer has that. So currently it's not possible for the base element queue to ever be used.

Awesomely, this takes along the concepts "sorted element queue", "custom element order", "processing base element queue", and "processing base element queue flag".