Strip down custom elements and SD specs
This makes the custom elements spec redirect to the HTML Standard, and removes all contents from the shadow DOM spec that has been upstreamed, while making it clearer that what remains is just the not-yet-fully-integrated parts.

It also attempts to reduce the red squigglies in the remaining parts of the shadow DOM spec, to the best of my ReSpec knowledge.